REFACTOR all convert test suites to test_convert.py module
